Fix sign and conversion warnings (major)
This makes all integer type conversions that have potential data loss
explicit with calls that do range checks and raise an exception. After
this commit, qpdf builds with no warnings when -Wsign-conversion
-Wconversion is used with gcc or clang or when -W3 -Wd4800 is used
with MSVC. This significantly reduces the likelihood of potential
crashes from bogus integer values.

There are some parts of the code that take int when they should take
size_t or an offset. Such places would make qpdf not support files
with more than 2^31 of something that usually wouldn't be so large. In
the event that such a file shows up and is valid, at least qpdf would
raise an error in the right spot so the issue could be legitimately
addressed rather than failing in some weird way because of a silent
overflow condition.

+++ libqpdf/InsecureRandomDataProvider.cc
@@ -30,7 +30,8 @@ InsecureRandomDataProvider::random()
 	// Seed the random number generator with something simple, but
 	// just to be interesting, don't use the unmodified current
 	// time.  It would be better if this were a more secure seed.
-        QUtil::srandom(QUtil::get_current_time() ^ 0xcccc);
+        QUtil::srandom(static_cast<unsigned int>(
+                           QUtil::get_current_time() ^ 0xcccc));
 	this->seeded_random = true;
     }
